requirements.txt
discord.py==1.7.1
dnspython==2.1.0
PyNaCl==1.4.0
async-timeout==3.0.1
NHentai-API==0.0.16
requests==2.25.1
aiohttp==3.7.4
python-dotenv==0.17.0
jishaku==2.0.0
jikanpy==4.3.1
rule34==1.8.1
#pyosu==1.1
firebase_admin==5.0.1
datetime==4.3